src/main/java/ru/netology/manager/MovieManager.java
package ru.netology.manager;

public class MovieManager {
    private final int limit;
    String[] movies = new String[0];

    public MovieManager(int limit) {
        this.limit = limit;
    }

    public MovieManager() {
        limit = 10;
    }

    public void add(String movie) {
        String[] tmp = new String[movies.length + 1];
        System.arraycopy(movies, 0, tmp, 0, movies.length);
        tmp[tmp.length - 1] = movie;

        movies = tmp;
    }

    public String[] findAll() {
        return movies;
    }

    public String[] findLast() {
        int resultLength = Math.min(movies.length, limit);
        String[] result = new String[resultLength];
        for (int i = 0; i < result.length; i++) {
            int index = movies.length - i - 1;
            result[i] = movies[index];
        }
        return result;
    }
}
